Testy vlozeny do souboru test_soustava.py
